Add CSV option for logging
Added a LOG_METHOD option to the configuration file to specify
a format to send to the syslog server.

Added the log_to_csv function - will create a csv string to log
to the syslog server.

Moved logic for building the CEF log to the log_to_cef function.

Created new function called "log_event" to log a specific with a
method based on the LOG_METHOD selected in the configuration.